move SearchService and PlaceCommentService
