--- core/src/main/java/greencity/dto/specification/SpecificationDto.java
@@ -1,13 +1,0 @@
-package greencity.dto.specification;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class SpecificationDto {
-    private Long id;
-    private String name;
-}
