--- core/src/main/java/greencity/dto/specification/SpecificationValueDto.java
@@ -1,14 +1,0 @@
-package greencity.dto.specification;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class SpecificationValueDto {
-    private Long id;
-    private String value;
-    private SpecificationDto specification;
-}
