+++ service-api/src/main/java/greencity/dto/photo/PhotoReturnDto.java
