+++ service-api/src/main/java/greencity/service/SearchService.java
